# kenyan-real-estate-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_verified:
        type: boolean
      last_name:
        type: string
      phone_number:
        type: string
      profile_image_url:
        type: string
      updated_at:
        type: string
      user_type:
        $ref: '#/definitions/models.UserType'
    type: object
  models.UserType:
    enum:
    - landlord
    - tenant
    - agent
    type: string
    x-enum-varnames:
    - UserTypeLandlord
    - UserTypeTenant
    - UserTypeAgent
...
  models.VerifyEmailRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive API for managing real estate properties in Kenya
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kenyan Real Estate API
  version: "1.0"
paths:
  /amenities:
    get:
      consumes:
      - application/json
      description: Get all available property amenities, optionally filtered by category
      parameters:
      - description: Filter by category
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of amenities
          schema:
            properties:
              amenities:
                type: object
            type: object
      summary: Get property amenities
      tags:
      - Kenyan Features
...
  /format-currency:
    post:
      consumes:
      - application/json
      description: Format an amount in Kenyan Shillings (KES)
      parameters:
      - description: Amount to format
        in: body
        name: amount_data
        required: true
        schema:
          properties:
            amount:
              type: number
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Formatted currency
          schema:
            properties:
              amount:
                type: number
              formatted:
                type: string
            type: object
        "400":
          description: Invalid request data
          schema:
            properties:
              details:
                type: string
              error:
                type: string
            type: object
      summary: Format currency
      tags:
      - Kenyan Features
...
  /popular-areas:
    get:
      consumes:
      - application/json
      description: Get popular residential areas, optionally filtered by county
      parameters:
      - description: Filter by county name
        in: query
        name: county
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of popular areas
          schema:
            properties:
              areas:
                type: object
            type: object
        "404":
          description: County not found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get popular areas
      tags:
      - Kenyan Features
...
  /send-verification-email:
    post:
      consumes:
      - application/json
      description: Send a verification email to the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Verification email sent
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: User already verified
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "429":
          description: Too many requests
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Send email verification
      tags:
      - Email Verification
...
  /validate-phone:
    post:
      consumes:
      - application/json
      description: Validate a Kenyan phone number format
      parameters:
      - description: Phone number to validate
        in: body
        name: phone_data
        required: true
        schema:
          properties:
            phone_number:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Validation result
          schema:
            properties:
              is_valid:
                type: boolean
              phone_number:
                type: string
            type: object
        "400":
          description: Invalid request data
          schema:
            properties:
              details:
                type: string
              error:
                type: string
            type: object
      summary: Validate phone number
      tags:
      - Kenyan Features
  /verification-status:
    get:
      consumes:
      - application/json
      description: Get the email verification status for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Verification status
          schema:
            properties:
              can_resend:
                type: boolean
              is_verified:
                type: boolean
              pending_verification:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - Bearer: []
      summary: Get verification status
      tags:
      - Email Verification
  /verify-email:
    post:
      consumes:
      - application/json
      description: Verify email address using the verification token
      parameters:
      - description: Verification token
        in: body
        name: verification
        required: true
        schema:
          $ref: '#/definitions/models.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Email verified successfully
          schema:
            properties:
              message:
                type: string
              user:
                $ref: '#/definitions/models.UserResponse'
            type: object
        "400":
          description: Invalid token or expired
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Token not found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Verify email address
      tags:
      - Email Verification
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"